src/domain/forum/application/use-cases/fetch-question-comments.ts
import { Either, right } from '@/core/either';
import { Injectable } from '@nestjs/common';
import { CommentWithAuthor } from '../../enterprise/entities/value-object/comment-with-author';
import { QuestionCommentsRepository } from '../repositories/question-comments-repository';

type Input = {
  questionId: string;
  page: number;
};

type Output = Either<
  null,
  {
    comments: CommentWithAuthor[];
  }
>;

@Injectable()
export class FetQuestionCommentUseCase {
  constructor(
    private readonly questionCommentsRepository: QuestionCommentsRepository,
  ) {}

  async execute({ questionId, page }: Input): Promise<Output> {
    const comments =
      await this.questionCommentsRepository.findManyByQuestionIdWithAuthor(
        questionId,
        {
          page,
        },
      );

    return right({
      comments,
    });
  }
}
